requirements.txt
Django>=1.11
djangorestframework
markdown
django-filter
psycopg2
